requirements.txt
absl-py==0.11.0
addict==2.4.0
alabaster==0.7.12
anyio==3.5.0
argon2-cffi==20.1.0
asn1crypto==1.2.0
astroid==2.3.3
attrs==19.3.0
Babel==2.7.0
backcall==0.1.0
bleach==3.1.0
cachetools==4.2.0
certifi==2021.10.8
cffi==1.13.2
chardet==3.0.4
charset-normalizer==2.0.12
cloudpickle==1.2.2
colorama==0.4.1
cryptography==2.8
cycler==0.10.0
Cython==0.29.28
decorator==4.4.1
defusedxml==0.6.0
docutils==0.15.2
dominate==2.7.0
editdistance==0.6.0
entrypoints==0.3
future==0.18.2
google-auth==1.24.0
google-auth-oauthlib==0.4.2
grpcio==1.34.0
h11==0.12.0
h5py==2.10.0
httpcore==0.14.7
httpx==0.22.0
idna==2.8
imageio==2.16.1
imagesize==1.1.0
importlib-metadata==0.23
ipykernel==5.1.3
ipython==7.9.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.15.1
Jinja2==2.10.3
joblib==0.14.0
jsonschema==3.2.0
jupyter-client==5.3.4
jupyter-core==4.6.1
keyring==19.2.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
lxml==4.6.3
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.4.3
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
mmcv==1.3.1
more-itertools==7.2.0
nbconvert==5.6.1
nbformat==4.4.0
notebook==6.1.5
numpy==1.21.2
numpydoc==0.9.1
oauthlib==3.1.0
olefile==0.46
opencv-python==4.5.1.48
orjson==3.6.7
packaging==19.2
pandas==0.25.3
pandocfilters==1.4.2
parso==0.5.1
pickleshare==0.7.5
Pillow==9.1.0
polygon==1.0.5
Polygon3==3.0.9.1
prometheus-client==0.9.0
prompt-toolkit==2.0.10
protobuf==3.14.0
psutil==5.6.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyclipper==1.3.0.post2
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
Pygments==2.4.2
pylint==2.4.4
pyOpenSSL==19.1.0
pyparsing==2.4.5
pyrsistent==0.15.6
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2019.3
pywin32==227
pywin32-ctypes==0.2.0
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq==18.1.0
QtAwesome==0.6.0
qtconsole==4.6.0
QtPy==1.9.0
regex==2022.3.2
requests==2.22.0
requests-oauthlib==1.3.0
rfc3986==1.5.0
rope==0.14.0
rsa==4.6
scikit-learn==0.22
scikit-video==1.1.11
scipy==1.7.1
seaborn==0.11.2
Send2Trash==1.5.0
six==1.13.0
sklearn==0.0
sniffio==1.2.0
snowballstemmer==2.0.0
Sphinx==2.2.1
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
spyder==3.3.6
spyder-kernels==0.5.2
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.7.0
terminado==0.9.1
testpath==0.4.4
thop==0.0.31.post2005241907
torch==1.9.0
torchaudio==0.9.0
torchlight==1.0
torchsummary==1.5.1
torchvision==0.10.0+cu111
tornado==6.0.3
tqdm==4.61.2
traitlets==4.3.3
typing-extensions==3.7.4.3
urllib3==1.24.2
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==1.3.1
websockets==10.2
Werkzeug==1.0.1
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.11.2
xlrd==1.2.0
yapf==0.32.0
zipp==0.6.0
